refactor: :recycle: styles files refactor and hooks refactor
