+++ src/components/Breadcrumb/styles.ts
@@ -1,0 +1,5 @@
+import { css } from "@emotion/react";
+
+export const breadcrumbContent = (theme) => css(({padding: '1px 24px 0', backgroundColor: theme?.colors?.white,}))
+
+export const breadcrumbStyles = css({ marginTop: '16px' })
