+++ src/components/JsonSettings/styles.ts
@@ -1,0 +1,11 @@
+import { css } from "@emotion/react"
+
+export const settingsContent = css({
+  display: "flex",
+  flexDirection: "column",
+  gap: "32px"
+})
+
+export const widthStyles = css({ width: "100%" })
+
+export const textInputsStyles = css({ color: "rgba(0, 0, 0, 0.25)", ":focus": {color: "#000000"}})
