+++ src/components/JsonView/styles.ts
@@ -1,0 +1,3 @@
+import { css } from "@emotion/react"
+
+export const widthStyles = css({ width: "100%" })
