+++ src/components/MainPreview/styles.ts
@@ -1,0 +1,5 @@
+import { css } from "@emotion/react";
+
+export const previewContent = css({ paddingLeft: "19px", width: "40%" });
+
+export const cardPreview = css({ height: "100%" })
