+++ src/components/SidebarFlow/hook.ts
@@ -1,0 +1,21 @@
+import { useState } from "react";
+import * as M from "@/components/Flow/mocks/mockTracks";
+
+
+const useSidebarFlow = () => {
+  const [optionsFlow, setOptionsFlow] = useState(M.optionsFlow);
+  
+  const addOptions = () =>
+    setOptionsFlow([
+      ...optionsFlow,
+      `Flow Tracks ${optionsFlow.length + 1}`,
+    ])
+
+  return {
+    optionsFlow,
+    addOptions
+  }
+  
+}
+
+export default useSidebarFlow
