+++ src/components/SidebarFlow/styles.ts
@@ -1,0 +1,17 @@
+import { css } from "@emotion/react";
+
+export const contentSidebarStyles = css({ paddingRight: "19px" })
+
+export const sidebarTitleStyles = css({
+  padding: "16px 24px",
+  margin: "0",
+  backgroundColor: "#fff",
+  fontWeight: "500",
+  fontSize: "16px",
+})
+
+export const dividerStyles = css({ margin: "0" })
+
+export const menuContentStyles = css({ width: "240px",  height: `${document.body.clientHeight/ 1.3}px` })
+
+export const addLinkStyles = css({color: "#148EFF" })
