+++ src/components/StepFields/styles.ts
@@ -1,0 +1,11 @@
+import { css } from "@emotion/react";
+
+export const widthStyles = css({ width: "100%" })
+
+export const contentStyles = css({ width: "60%" })
+
+export const deleteButton = css({background: "none", border: "none"})
+
+export const emailInputContent = css({ width: "100%", gap: "16px", display: "flex" })
+
+export const emailInput = css({ width: "49%", display: "flex", flexDirection: "column", gap: "8px" })
