+++ src/components/ValidatorsModal/components/CollapseHeader/styles.ts
@@ -1,17 +1,19 @@
-export const buttonsStyles = {
+import { css } from "@emotion/react"
+
+export const buttonsStyles = css({
 	height: "fit-content",
 	padding: "0",
 	marginLeft: "10px"
-}
+})
 
-export const textStyles = {
+export const textStyles = css({
 	margin: "0"
-}
+})
 
-export const headerStyles = {
+export const headerStyles = css({
 	display: "flex",
 	alignItems: "center",
 	justifyContent: "space-between",
 	width: "100%",
-}
+})
