+++ src/components/ValidatorsModal/components/CollapseProperties/styles.ts
@@ -1,3 +1,5 @@
-export const inputGroupStyles = {
+import { css } from "@emotion/react"
+
+export const inputGroupStyles = css({
 	marginTop: "25px"
-}
+})
