+++ src/components/ValidatorsModal/components/Properties/styles.ts
@@ -1,4 +1,6 @@
-export const buttonsStyles = {
+import { css } from "@emotion/react";
+
+export const buttonsStyles = css({
 	height: "fit-content",
 	padding: "0"
-}
+})
