+++ src/pages/Editor/JSONTab/styles.ts
@@ -1,4 +1,4 @@
-import { css } from "@emotion/react";
+import { css } from "@emotion/react"
 
 export const contentStyles = css({
   padding: "0 28px 28px 28px",
@@ -6,14 +6,4 @@ export const contentStyles = css({
   display: "flex",
   width: "100%",
   gap: "29px"
-});
-
-export const settingsContent = css({
-  display: "flex",
-  flexDirection: "column",
-  gap: "32px"
 })
-
-export const widthStyles = { width: "100%" };
-
-export const textInputsStyles = { color: "rgba(0, 0, 0, 0.25)", ":focus": {color: "#000000"}  };
