+++ src/pages/Editor/TracksTab/styles.ts
@@ -1,22 +1,8 @@
-export const contentStyles: any = {
+import { css } from "@emotion/react";
+
+export const contentStyles: any = css({
   padding: "28px",
   paddingTop: 0,
   flexDirection: "row",
   justifyContent: "center",
-};
-
-export const contentSidebarStyles = { paddingRight: "19px" };
-
-export const sidebarTitleStyles = {
-  padding: "16px 24px",
-  margin: "0",
-  backgroundColor: "#fff",
-  fontWeight: "500",
-  fontSize: "16px",
-};
-
-export const dividerStyles = { margin: "0" };
-
-export const menuContentStyles = { width: "240px",  height: `${document.body.clientHeight/ 1.3}px` };
-
-export const addLinkStyles = {color: "#148EFF" };
+})
