+++ src/pages/Editor/index.tsx
@@ -1,10 +1,8 @@
-// Todo - separar em arquivos menores p/ components
-import React from "react";
 import { Outlet, useNavigate, useLocation } from "react-router-dom";
+import BreadcrumbComponent from "@/components/Breadcrumb";
+import { tabsMenu } from "@/components/TabsMenu";
 import {
   Layout,
-  Menu,
-  Breadcrumb,
   PageHeader,
   Button,
   Tabs,
@@ -15,35 +13,10 @@ import {
   SisternodeOutlined,
   FileOutlined,
   EllipsisOutlined,
-  DownloadOutlined,
 } from "@ant-design/icons";
 import * as S from "./styles";
 
-const { Content } = Layout;
 const { TabPane } = Tabs;
-
-const tabsMenu = (
-  <Menu>
-    <Menu.Item key="1" icon={<DownloadOutlined />}>
-      <a
-        // target="_blank"
-        rel="noopener noreferrer"
-        href="/"
-      >
-        Export form JSON
-      </a>
-    </Menu.Item>
-  </Menu>
-);
-
-const BreadcrumbComponent = () => (
-  <div css={S.breadcrumbContent}>
-    <Breadcrumb css={S.breadcrumbStyles}>
-      <Breadcrumb.Item>Experiences</Breadcrumb.Item>
-      <Breadcrumb.Item>My Form</Breadcrumb.Item>
-    </Breadcrumb>
-  </div>
-);
 
 
 const tabs = [
@@ -61,7 +34,6 @@ const EditorPage = () => {
   return (
     <>
       <Layout className="layout" css={S.layoutStyles}>
-        <Content css={S.contentStyles}>
           <BreadcrumbComponent />
           <PageHeader
             css={S.pageHeaderStyles}
@@ -98,7 +70,6 @@ const EditorPage = () => {
               ))}
             </Tabs>
           </PageHeader>
-        </Content>
       </Layout>
       <Outlet />
     </>
