+++ src/pages/Editor/styles.ts
@@ -1,18 +1,7 @@
-export const layoutStyles = { padding: '28px'}
+import { css } from "@emotion/react";
 
-export const contentStyles = { 
-    // padding: '50px 28px'
-}
+export const layoutStyles = css({ padding: '28px'})
 
-// export const layoutStyles = { padding: '28px 0 0', display: "flex", alignItems: "center"} // todo
+export const pageHeaderStyles = css({padding: "16px 24px 0"})
 
-// export const contentStyles = { padding: '50px 28px 22px', width: "1452px"} // todo
-
-
-export const pageHeaderStyles = {padding: "16px 24px 0"}
-
-export const breadcrumbContent = (theme) => ({padding: '1px 24px 0', backgroundColor: theme?.colors?.white,})
-
-export const breadcrumbStyles = { marginTop: '16px' }
-
-export const moreButton = (theme) => ({borderColor: theme?.colors?.["grey85"]})
+export const moreButton = (theme) => (css({borderColor: theme?.colors?.["grey85"]}))
